feat: Add hook to manage current folder
